Add local user caching system
The reason I added this is because I couldn't figure out how to wait
for all the comment's content & user details to fully load before
displaying comments in a thread. There was a really ugly
"display as each comment loads" effect that is now sorted with the
caching. Also fixed a bug where a user would create a thread and be
sent to the wrong url